Add complicated pinning feature
This adds pinning functionality, by having extra-container/eval-config re-call itself when it detects `extra.overrideArgs` is set.
This allows, for example, to pin nixpkgs with niv, by overriding the nixosPath argument with an expression such as
 {
  ...
  extra = {
    overrideArgs = { nixosPath = (import ./sources.nix {}).nixpkgs + "/nixos"; };
    };
  ...
}

Note that extra-container technically allows specifiying multiple containers in a file, however the pinning functionality cannot currently handle this.
The settings in the first (by nix attrset ordering) container are used.

Changes:
I changed the function signature of eval-config to take an attrset. This lets us pass on the `args` alias.
I added an option to extraModule.
There is still some ad-hoc debugging with tracing, which is imported from lib and so depends on `<>/nixos` being in a nixpkgs checkout, so that lib can be imported.
If stage1.config.containers.<somecontainer>.extra.overrideArgs is set, a detour is taken through the stage2 function which does the actual overriding.